Extract message handling from index.js
